Update: Collapsible Drop Down + RaceCount hooks
